refactor(compose): split replydraft.go by concept under the file-length cap
PR #949 grew the file past the 500-line ceiling both go-file-length and
craft static enforce whole-tree, breaking every backend-touching PR. A
mechanical move, no behaviour change: the orchestrator stays in
replydraft.go, the Voice DNA lane moves to replydraftvoice.go, the model
request/validation lane to replydraftmodel.go.
