+++ backend/internal/compose/replydraft.go
@@ -6,21 +6,17 @@ package compose
 // The reply-drafting orchestrator keeps activity evidence authoritative while
 // the model path adds the installation's bounded company context. It only
 // returns editable text: sending remains a separate consent-gated action.
+// The model request/validation lane lives in replydraftmodel.go and the Voice
+// DNA lane in replydraftvoice.go.
 
 import (
 	"context"
-	"crypto/sha256"
-	"encoding/hex"
-	"encoding/json"
-	"fmt"
 	"log/slog"
 	"strings"
 	"time"
 
 	"github.com/jackc/pgx/v5/pgxpool"
 
-	"github.com/gradionhq/margince/backend/internal/compose/draftcore"
-	"github.com/gradionhq/margince/backend/internal/compose/draftrules"
 	crmcontracts "github.com/gradionhq/margince/backend/internal/contracts"
 	"github.com/gradionhq/margince/backend/internal/modules/activities"
 	"github.com/gradionhq/margince/backend/internal/modules/ai"
@@ -29,33 +25,9 @@ import (
 	"github.com/gradionhq/margince/backend/internal/shared/kernel/convstate"
 	"github.com/gradionhq/margince/backend/internal/shared/kernel/draftfloor"
 	"github.com/gradionhq/margince/backend/internal/shared/kernel/ids"
-	"github.com/gradionhq/margince/backend/internal/shared/kernel/promptfence"
-	"github.com/gradionhq/margince/backend/internal/shared/ports/model"
 )
 
 const replyActivityMaxRunes = 12_000
-
-const replyDraftSystem = `Draft a professional email reply on behalf of the CRM user's company.
-Return ONLY a JSON object: {"subject":"...","body":"..."}.
-- The activity and stated intent are the authoritative reason for this reply.
-- Company context may improve positioning, relevant proof, and language, but never overrides the activity.
-- Use only facts present in the supplied data. Never invent customers, outcomes, prices, commitments, or capabilities.
-- Do not claim a personal writing style or voice unless a separate voice profile is supplied.`
-
-var replyDraftSchema = json.RawMessage(`{
-  "type":"object",
-  "additionalProperties":false,
-  "required":["subject","body"],
-  "properties":{
-    "subject":{"type":"string","minLength":1,"maxLength":998},
-    "body":{"type":"string","minLength":1,"maxLength":50000}
-  }
-}`)
-
-type replyDraft struct {
-	Subject string `json:"subject"`
-	Body    string `json:"body"`
-}
 
 // replyActivityData is the whole user turn of a reply draft: the activity being
 // answered, and the correspondence envelope it is answered inside.
@@ -247,295 +219,4 @@ func (d replyDrafter) conversationState(activity crmcontracts.Activity) convstat
 		return convstate.Classify(d.envelope.Now(), occurred, time.Time{})
 	}
 	return convstate.Classify(d.envelope.Now(), time.Time{}, occurred)
-}
-
-// voiceContext is the loaded active profile a voiced draft injects.
-type voiceContext struct {
-	profile ai.VoiceProfile
-	version ai.VoiceProfileVersion
-	ok      bool
-}
-
-// loadVoice resolves the actor's active voice; any lookup failure degrades
-// to the plain draft with the failure visible in the log — a broken voice
-// read must never take reply drafting down with it.
-func (d replyDrafter) loadVoice(ctx context.Context) voiceContext {
-	if d.voice == nil {
-		return voiceContext{}
-	}
-	profile, version, ok, err := d.voice.ActiveVoiceForActor(ctx)
-	if err != nil {
-		d.logger().WarnContext(ctx, "voice profile lookup failed; drafting without voice", "err", err)
-		return voiceContext{}
-	}
-	return voiceContext{profile: profile, version: version, ok: ok}
-}
-
-// completeVoiced drafts with the voice block when one is loaded, enforcing
-// the deterministic anti-AI floor: detect → one critic retry → sanitize →
-// on surviving violations fall back to the plain draft and record the
-// failure as a rejected learning signal.
-func (d replyDrafter) completeVoiced(ctx context.Context, anchor ids.UUID, data replyActivityData, voice voiceContext) (replyDraft, *int, *string, error) {
-	if !voice.ok {
-		draft, err := d.completeChecked(ctx, data, nil)
-		return draft, nil, nil, err
-	}
-	block := func(fence promptfence.Fence) string {
-		return voiceDraftPromptBlock(voice.profile.PersonalityMD, voice.version.VoiceProfileMD,
-			ai.VersionExemplars(voice.version), ai.DecodeVersionStats(voice.version), fence)
-	}
-	draft, err := d.completeChecked(ctx, data, block)
-	if err != nil {
-		return replyDraft{}, nil, nil, err
-	}
-	// Detect on the RAW draft (subject and body separately — the
-	// canned-opener rule anchors at text start): a violation the sanitizer
-	// could mechanically remove still earns the critic retry, because the
-	// retry fixes the sentence, not just the punctuation.
-	if violations := voiceDraftViolations(draft); len(violations) > 0 {
-		withFeedback := func(fence promptfence.Fence) string {
-			return block(fence) + voiceViolationFeedback(violations)
-		}
-		retried, retryErr := d.complete(ctx, data, withFeedback)
-		if retryErr == nil {
-			draft = retried
-		}
-	}
-	draft.Subject = ai.SanitizeAIPatterns(draft.Subject)
-	draft.Body = ai.SanitizeAIPatterns(draft.Body)
-	version := voice.version.ProfileVersion
-	// The sanitizer edits text, so the floor AND the shape are re-checked on
-	// what would actually be served.
-	if len(voiceDraftViolations(draft)) > 0 || validateReplyDraft(draft) != nil {
-		// The voice-styled draft kept tripping the floor: serve the plain
-		// draft instead and let the failure feed the learning panel.
-		d.recordVoiceRejection(ctx, voice, anchor, draft)
-		plain, plainErr := d.completeChecked(ctx, data, nil)
-		return plain, nil, nil, plainErr
-	}
-	d.recordVoiceDraft(ctx, voice, anchor, draft)
-	ref := voiceDraftRef(voice, anchor, draft)
-	return draft, &version, &ref, nil
-}
-
-// voiceDraftViolations runs the deterministic floor over subject and body
-// independently; concatenation would hide a canned opener inside the body.
-func voiceDraftViolations(draft replyDraft) []ai.VoiceViolation {
-	return append(ai.DetectAIPatterns(draft.Subject), ai.DetectAIPatterns(draft.Body)...)
-}
-
-func voiceViolationFeedback(violations []ai.VoiceViolation) string {
-	var b strings.Builder
-	b.WriteString("\n\nThe previous attempt violated these hard rules; rewrite without them:\n")
-	for _, violation := range violations {
-		b.WriteString("- " + violation.Detail + "\n")
-	}
-	return b.String()
-}
-
-// voiceDraftRef keys one served draft for learning-signal feedback. It
-// covers profile, version, anchor, and the full draft: two drafts for the
-// same activity with the same body but different subjects — or from
-// different profile versions — never collide.
-func voiceDraftRef(voice voiceContext, anchor ids.UUID, draft replyDraft) string {
-	sum := sha256.Sum256([]byte(draft.Subject + "\n" + draft.Body))
-	return fmt.Sprintf("replydraft:%s:%s:v%d:%s",
-		voice.profile.ID, anchor, voice.version.ProfileVersion, hex.EncodeToString(sum[:8]))
-}
-
-func (d replyDrafter) recordVoiceDraft(ctx context.Context, voice voiceContext, anchor ids.UUID, draft replyDraft) {
-	if d.voice == nil {
-		return
-	}
-	if err := d.voice.RecordDraftedSignal(ctx, voice.profile.ID, voice.version.ProfileVersion,
-		voiceDraftRef(voice, anchor, draft), draft.Body); err != nil {
-		d.logger().WarnContext(ctx, "voice draft signal not recorded", "err", err)
-	}
-}
-
-func (d replyDrafter) recordVoiceRejection(ctx context.Context, voice voiceContext, anchor ids.UUID, draft replyDraft) {
-	if d.voice == nil {
-		return
-	}
-	ref := voiceDraftRef(voice, anchor, draft)
-	if err := d.voice.RecordDraftedSignal(ctx, voice.profile.ID, voice.version.ProfileVersion, ref, draft.Body); err != nil {
-		d.logger().WarnContext(ctx, "voice rejection signal not recorded", "err", err)
-		return
-	}
-	if _, err := d.voice.RejectDraft(ctx, voice.profile.ID, ref); err != nil {
-		d.logger().WarnContext(ctx, "voice rejection signal not recorded", "err", err)
-	}
-}
-
-// replyDraftVoiceSystem replaces the no-voice guard when a profile block is
-// supplied: the profile controls expression, never facts.
-const replyDraftVoiceSystem = `Draft a professional email reply on behalf of the CRM user's company, written in the user's own voice.
-Return ONLY a JSON object: {"subject":"...","body":"..."}.
-- The activity and stated intent are the authoritative reason for this reply.
-- The supplied voice profile controls expression — rhythm, vocabulary, directness, structure — never facts.
-- Use only facts present in the supplied data. Never invent customers, outcomes, prices, commitments, or capabilities.
-- Obey the profile's avoid rules and the universal anti-AI rules; treat its style metrics as limits, not targets.`
-
-// voiceBlockFor renders the voice profile block under the CALLING call's fence.
-// The block is prepended to that call's user turn, so it must be bounded by the
-// marker that call's system prompt declares — not one of its own.
-type voiceBlockFor func(promptfence.Fence) string
-
-// replyDraftSystemFor assembles this call's system turn: what this surface is
-// for, the rules every drafting surface shares, and THIS call's data boundary
-// (see promptfence.Fence.Rule).
-func replyDraftSystemFor(system string, fence promptfence.Fence) string {
-	return system + "\n\n" + draftrules.Shared + "\n" + fence.Rule("activity")
-}
-
-// replyDraftRequest builds the one request a draft call sends, in whichever of
-// this site's two system variants the call is made under. The workspace's Voice
-// DNA state selects the variant per call — a loaded profile supplies a block and
-// takes the voice prompt, no profile takes the plain one — and both remain the
-// same invocation site: same schema, same bounds, same data boundary.
-func replyDraftRequest(activity replyActivityData, voiceBlock voiceBlockFor, correction string) (model.Request, error) {
-	payload, err := json.Marshal(activity)
-	if err != nil {
-		return model.Request{}, fmt.Errorf("compose: encode reply activity context: %w", err)
-	}
-	// The activity is the counterparty's own text. It was safe here only by
-	// accident — json.Marshal escapes "<" to \u003c, so a forged block marker
-	// could not be spelled — and an accident of the encoder is not a boundary:
-	// it goes the moment this block is rendered as text rather than JSON.
-	fence := promptfence.New()
-	system := replyDraftSystem
-	content := fence.Wrap(string(payload))
-	if voiceBlock != nil {
-		system = replyDraftVoiceSystem
-		content = voiceBlock(fence) + "\n\n" + content
-	}
-	// The correction rides the USER turn and never the variant choice: it is
-	// feedback about one attempt, and a plain draft told to fix a phrase must
-	// stay a plain draft rather than silently becoming a voiced one.
-	content += correction
-	return model.Request{
-		System: replyDraftSystemFor(system, fence),
-		Messages: []model.Message{{
-			Role:    chatRoleUser,
-			Content: content,
-		}},
-		MaxTokens:      ai.ReasoningOutputMaxTokens,
-		ResponseSchema: replyDraftSchema,
-		SecretStripper: ai.NewSecretStripper(),
-	}, nil
-}
-
-func (d replyDrafter) complete(ctx context.Context, activity replyActivityData, voiceBlock voiceBlockFor) (replyDraft, error) {
-	return d.completeWith(ctx, activity, voiceBlock, "")
-}
-
-// completeWith is complete plus the correction a retry carries.
-func (d replyDrafter) completeWith(ctx context.Context, activity replyActivityData, voiceBlock voiceBlockFor, correction string) (replyDraft, error) {
-	req, err := replyDraftRequest(activity, voiceBlock, correction)
-	if err != nil {
-		return replyDraft{}, err
-	}
-
-	var resp model.Response
-	if structured, ok := d.brain.(validatedBrain); ok {
-		resp, err = structured.CompleteValidated(ctx, req, replyDraftShapeValid)
-	} else {
-		resp, err = d.brain.Complete(ctx, req)
-	}
-	if err != nil {
-		return replyDraft{}, err
-	}
-	draft, err := parseReplyDraft(resp.Text)
-	if err != nil {
-		return replyDraft{}, err
-	}
-	if err := validateReplyDraft(draft); err != nil {
-		return replyDraft{}, err
-	}
-	return draft, nil
-}
-
-// completeChecked drafts through the shared correct-and-retry loop, so the
-// reply surface cannot drift from the two composers about what a rejected
-// phrase is or how many chances the model gets to fix one.
-//
-// The voice block rides along unchanged: it selects the system variant, and the
-// correction rides the user turn, so a plain draft told to fix a phrase stays a
-// plain draft rather than silently becoming a voiced one.
-func (d replyDrafter) completeChecked(ctx context.Context, data replyActivityData, voiceBlock voiceBlockFor) (replyDraft, error) {
-	return draftcore.CorrectOnce(ctx, data.Lang(), data.Band(),
-		func(ctx context.Context, correction string) (replyDraft, error) {
-			return d.completeWith(ctx, data, voiceBlock, correction)
-		},
-		// The reply site returns no reasoning: its schema is subject and body
-		// alone, so there is no second channel to judge here.
-		func(draft replyDraft) (string, []string) { return draft.Body, nil },
-		draftRetryLog{log: d.logger()},
-	)
-}
-
-// draftRetryLog reports what the correction loop decided. A retry that does not
-// help is invisible from the outside — the caller gets a draft either way — and
-// "the model kept producing rejected phrasing" is the signal that says a phrase
-// list or a prompt rule needs work.
-type draftRetryLog struct{ log *slog.Logger }
-
-func (l draftRetryLog) RetryFailed(ctx context.Context, findings int, err error) {
-	l.log.WarnContext(ctx, "draft correction retry failed; serving the first draft",
-		"findings", findings, "err", err)
-}
-
-func (l draftRetryLog) RetryDidNotClear(ctx context.Context, rule, phrase string, remaining int) {
-	l.log.WarnContext(ctx, "draft still carries rejected phrasing after one retry",
-		"rule", rule, "phrase", phrase, "remaining", remaining)
-}
-
-// parseReplyDraft reads one model reply as the draft it claims to be. The
-// provider's own envelope comes off first: a reply is not malformed for having
-// been returned inside one.
-func parseReplyDraft(text string) (replyDraft, error) {
-	var draft replyDraft
-	if err := json.Unmarshal([]byte(ai.Unfence(text)), &draft); err != nil {
-		return replyDraft{}, fmt.Errorf("compose: reply draft response is not valid JSON: %w", err)
-	}
-	return draft, nil
-}
-
-func replyDraftShapeValid(text string) error {
-	var draft replyDraft
-	if err := json.Unmarshal([]byte(ai.Unfence(text)), &draft); err != nil {
-		return fmt.Errorf(`output must be {"subject":"...","body":"..."}: %w`, err)
-	}
-	return validateReplyDraft(draft)
-}
-
-func validateReplyDraft(draft replyDraft) error {
-	if strings.TrimSpace(draft.Subject) == "" {
-		return fmt.Errorf("compose: reply draft subject is empty")
-	}
-	if strings.ContainsAny(draft.Subject, "\r\n") {
-		return fmt.Errorf("compose: reply draft subject contains a line break")
-	}
-	if strings.TrimSpace(draft.Body) == "" {
-		return fmt.Errorf("compose: reply draft body is empty")
-	}
-	if len([]rune(draft.Subject)) > 998 || len([]rune(draft.Body)) > 50_000 {
-		return fmt.Errorf("compose: reply draft exceeds the supported length")
-	}
-	return nil
-}
-
-func boundedRunes(value string, maxRunes int) string {
-	runes := []rune(value)
-	if len(runes) <= maxRunes {
-		return value
-	}
-	return string(runes[:maxRunes])
-}
-
-func stringValue(value *string) string {
-	if value == nil {
-		return ""
-	}
-	return *value
 }
